Update algebra solver 5.5.py
This is a python program based on either evaluating the exact value of an expression, or solving an equation for "x". It does not support answers given in terms of a variable. The main algorithm uses a searching method to find the exact value of "x", narrowing it's search every time it gets closer to the answer. I have a max iterations part that does not let you crash python looking for an imaginary number. If the program never finds it's first clue within the max iterations, it will stop itself. This does not use any external libraries other than the ones that come with python (parser, tkinter, and math). Written in python 3.8.1 on 4/20/2020